moved imagecomponent code to projectbarcomponent
